+++ src/stories/2-configure-columns.story.vue
@@ -39,6 +39,21 @@
         }"
       />
     </Variant>
+    <Variant title="Extra columns" width="100">
+      <VueScreener
+        :data="baseData"
+        :columns="{
+          id: { field: 'id', width: '50px' },
+          first_name: { field: 'first_name', width: '150px' },
+          last_name: { field: 'last_name', width: '150px' },
+          full_name: { field: 'full_name', width: '150px' },
+          email: { field: 'email', width: '200px' },
+          gender: { field: 'gender', width: '100px' },
+          ip_address: { field: 'ip_address', width: 'minmax(150px, 1fr)' },
+          extra_column: { with: '100px' },
+        }"
+      />
+    </Variant>
     <Variant title="Pin to the side" width="100">
       <VueScreener :data="baseData" include-pinned>
         <template #pinned-head="props">
